backend_django/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.0.1
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==20.0.0
numpy==1.24.3
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.0
whitenoise==6.6.0
python-dotenv==1.0.0
dj-database-url==1.3.0